requirements.txt
streamlit
opencv-python
scikit-image
pillow
numpy